+++ data/users.js
@@ -25,7 +25,6 @@ export async function checkUsernameAndEmail(username, email) {
 
 // create user
 export async function createUser(username, password, email) {
-  // validateUser(username, password, email, pfp, lastfm);
   username = validate.validName(username);
   password = validate.validPassword(password);
   email = validate.validEmail(email);
@@ -36,7 +35,7 @@ export async function createUser(username, password, email) {
   await checkUsernameAndEmail(username, email);
 
   //encrypt password
-  const hash = await bcrypt.hash(password, 16);
+  const hash = await bcrypt.hash(password, 4); //remember to change to back to 16 or 12 for all bcrypts
 
   const pfp = 'https://source.unsplash.com/1600x900/?' + username;
 
@@ -60,12 +59,13 @@ export async function createUser(username, password, email) {
   if (!insertInfo.acknowledged || !insertInfo.insertedId) throw "Could not add user";  
   const newId = insertInfo.insertedId.toString();
   const user = await getUserById(newId);
+  if (!user) throw "User not found";
   return user;
 }
 
 // get user by id
 export async function getUserById(id) {
-  // handleId(id);
+  id = validate.validId(id);
   const userCollection = await users();
   const user = await userCollection.findOne({ _id: new ObjectId(id) });
   if (!user) throw "User not found";
@@ -105,9 +105,10 @@ export async function getAllUsers() {
 
 // remove user by id
 export async function removeUserById(id) {
-  // handleId(id);
+  id = validate.validId(id);
   const userCollection = await users();
   const user = await getUserById(id);
+  if (!user) throw `Could not find user with id of ${id}`;
   const deletionInfo = await userCollection.deleteOne({ _id: ObjectId(id) });
   if (!deletionInfo.acknowledged || deletionInfo.deletedCount === 0) throw `Could not delete user with id of ${id}`;
   return user;
@@ -115,14 +116,17 @@ export async function removeUserById(id) {
 
 // update user by id
 export async function updateUserById(id, username, password, lastfmUsername) {
-  // todo
-  // handleId(id);
-  
-  // validateUser(updatedUser);
+
+  id = validate.validId(id);
+  username = validate.validEditedUsername(username);
+  password = validate.validEditedPassword(password);
+  lastfmUsername = validate.validFmString(lastfmUsername);
+
   const userCollection = await users();
   const user = await getUserById(id);
+  if (!user) throw `Could not find user with id of ${id}`;
   const lastfmData = lastfmUsername ? await lastfm.getInfoByUser(lastfmUsername) : null;
-  let hash = (password === '') ? null : await bcrypt.hash(password, 10); //if password is empty string, dont update password
+  let hash = (password == null) ? null : await bcrypt.hash(password, 4);
 
   const updatedUser = {
     username: username || user.username,
@@ -139,6 +143,14 @@ export async function updateUserById(id, username, password, lastfmUsername) {
     updatedAt: new Date()
   };
 
+  // there might be a better way to do this
+  const updateInfo = await userCollection.findOneAndReplace(
+    { _id: new ObjectId(id) }, 
+    updatedUser,
+    { returnDocument: 'after' }
+    );
+  if (!updateInfo) throw `Error: Update failed! Could not update user with id of ${id}`;
+
   // need to update all the posts that the user has created
   // need to update all the posts that the user has liked
   // ? need to update all the posts that the user has commented on -- depending on how comments work
@@ -147,21 +159,13 @@ export async function updateUserById(id, username, password, lastfmUsername) {
   const postCollection = await posts();
   for (let i = 0; i < user.createdPosts.length; i++) {
     const post = await postCollection.findOneAndUpdate(
-      { _id: user.createdPosts[i]}, 
-      { $set: { userId: id,
-        username: updatedUser.username}},
+      { _id: new ObjectId(user.createdPosts[i])}, 
+      { $set: { username: updatedUser.username}},
       { returnDocument: 'after' }
     );
     if (!post) throw `Error: Update failed! Could not update post with id of ${user.createdPosts[i]}`;
   };
 
-  // there might be a better way to do this
-  const updateInfo = await userCollection.findOneAndReplace(
-    { _id: new ObjectId(id) }, 
-    updatedUser,
-    { returnDocument: 'after' }
-    );
-  if (!updateInfo) throw `Error: Update failed! Could not update user with id of ${id}`;
   return updateInfo;
 }
 
@@ -182,9 +186,13 @@ export const loginUser = async (email, password) => {
 export const followUser = async (userId, profileId) => { //adds profile to user following list and adds user to profile's followers list
   // handleId(followerId); 
   // handleId(followingId);
+  userId = validate.validId(userId);
+  profileId = validate.validId(profileId);
   const userCollection = await users();
   const user = await getUserById(userId);
+  if (!user) throw "User not found";
   const profile = await getUserById(profileId);
+  if (!profile) throw "Profile not found";
   if(user.following.includes(profileId)) throw "Already following user 1";
   if(profile.followers.includes(userId)) throw "Already following user 2";
   
@@ -210,9 +218,13 @@ export const followUser = async (userId, profileId) => { //adds profile to user 
 export const unfollowUser = async (userId, profileId) => { //removes profile form user following list and removes user form profile's followers list
   // handleId(followerId);
   // handleId(followingId);
+  userId = validate.validId(userId);
+  profileId = validate.validId(profileId);
   const userCollection = await users();
   const user = await getUserById(userId);
+  if (!user) throw "User not found";
   const profile = await getUserById(profileId);
+  if (!profile) throw "Profile not found";
   if (!user.following.includes(profileId)) throw "Not following user 1";
   if (!profile.followers.includes(userId)) throw "Not following user 2";
 
@@ -236,13 +248,16 @@ export const unfollowUser = async (userId, profileId) => { //removes profile for
 
 export const addNotification = async (profileId, notification) => {
   // handleId(userId);
+  profileId = validate.validId(profileId);
+  notification = validate.validString(notification);
   const userCollection = await users();
-  const user = await getUserById(profileId);
+
   let newNotification = {
-    _id : new ObjectId(),
-    notification : notification,
-    dateCreated : new Date()
+    _id: new ObjectId().toString(),
+    notification: notification,
+    dateCreated: new Date()
   }
+  
   const insertNotification = await userCollection.findOneAndUpdate(
     { _id: new ObjectId(profileId) },
     { $push: { notifications: newNotification } },
@@ -250,4 +265,19 @@ export const addNotification = async (profileId, notification) => {
   );
   if (!insertNotification) throw "Error: Update failed! Could not add notification";
   return insertNotification;
+}
+
+export const removeNotification = async (userId, notificationId) => { //idk if this works
+  // handleId(userId);
+  // handleId(notificationId);
+  userId = validate.validId(userId);
+  notificationId = validate.validId(notificationId);
+  const userCollection = await users();
+  const removeNotification = await userCollection.findOneAndUpdate(
+    { _id: new ObjectId(userId) },
+    { $pull: { notifications: {_id : notificationId}} },
+    { returnDocument: 'after' }
+  );
+  if (!removeNotification) throw "Error: Update failed! Could not remove notification";
+  return removeNotification;
 }
